% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_ratings.R
\name{historical_elo_ratings}
\alias{historical_elo_ratings}
\title{Load Historical IndyCar Elo Ratings}
\usage{
historical_elo_ratings()
}
\description{
This function loads in historical IndyCar Elo ratings.
}
