% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_championship_points}
\alias{plot_championship_points}
\title{Plot the season-long championship points standings}
\usage{
plot_championship_points(season = 2020)
}
\description{
Plot the season-long championship points standings
}
