% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_results.R
\name{race_results}
\alias{race_results}
\title{Load IndyCar Race Results}
\usage{
race_results(season = 2020, race_number = 1)
}
\description{
This function loads in results and stats from IndyCar races.
}
