% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_stats.R
\name{season_stats}
\alias{season_stats}
\title{Calculate IndyCar Full Season Stats}
\usage{
season_stats(season = 2020, track_type = "all")
}
\description{
This function calculates full season basic and advanced IndyCar stats.
Advanced stats are not available for all seasons, so some columns might be
'NA' or have default values in them.
}
